refactor(di): move Dependency from key.js to binding.js
