+++ modules/di/src/key.js
@@ -4,18 +4,6 @@ import {FIELD, int, bool} from 'facade/lang';
 var _allKeys = {};
 var _id:int = 0;
 
-//TODO: vsavkin: move to binding once cyclic deps are supported
-@FIELD('final key:Key')
-@FIELD('final asFuture:bool')
-@FIELD('final lazy:bool')
-export class Dependency {
-  constructor(key:Key, asFuture:bool, lazy:bool) {
-    this.key = key;
-    this.asFuture = asFuture;
-    this.lazy = lazy;
-  }
-}
-
 @FIELD('final token')
 @FIELD('final id:int')
 export class Key {
